requirements.txt
Django==1.3.1
Pillow==1.7.6
django-endless-pagination==1.1
django-grappelli==2.3.7
gunicorn==0.13.4
sorl-thumbnail==11.12
textile==2.1.5
wsgiref==0.1.2
